Add check for template

+++ startday.sh
@@ -7,6 +7,8 @@ day=$1
 [[ "$day" =~ ^[0-9]{1,2}$ && $day -ge 1 && $day -le 31 ]] || { echo "'$day' is not a valid day"; usage; }
 shift
 
+# language
+
 language=""
 
 while getopts "l:" opt; do
@@ -20,13 +22,10 @@ case $language in
   "go" ) echo "Setting up day for Go" ;;
   "ipynb" ) echo "Setting up day for Jupyter Notebook" ;;
   "" ) { echo "No language chosen, proceeding with Go"; language="go"; } ;;
-  * ) { echo "'$language' is not a valid programming language. If you have a template for it, add it to the case in this script."; usage; }
+  * ) { echo "'$language' is not a valid programming language. If you have a template for it, add the language to the case in this script."; usage; }
 esac
 
-aoc="https://adventofcode.com/$year/day/$day"
-echo "Preparing Advent of Code $year day $day: $aoc"
-
-dir="$year/day$day"
+# Session cookie
 
 if [ -f .env ]; then
   export "$(xargs < .env)"
@@ -37,11 +36,20 @@ if [ "$AOC_SESSION" = "" ]; then
   exit 1
 fi
 
+# creating directory
+
+aoc="https://adventofcode.com/$year/day/$day"
+echo "Preparing Advent of Code $year day $day: $aoc"
+
+dir="$year/day$day"
+
 if test -d "$dir"; then
   echo "Directory ./$dir already exists"
   exit 0
 fi
 mkdir -p "$dir"
+
+# input.txt
 
 echo "Fetching $aoc/input"
 
@@ -52,6 +60,13 @@ curl -fsS --request GET "$aoc/input" -H "cookie: session=$AOC_SESSION" -o "$dir/
 }
 
 echo "Created $dir/index.txt"
+
+# main script
+
+if ! test -f "template/$language"; then
+  echo "Template template/$language does not exist"
+  exit 1
+fi
 
 cp "template/$language" "$dir/main.$language"
 echo "Created $dir/main.$language"
